+++ index.js
@@ -8,6 +8,46 @@ const { setSvgBbox, validSVG } = require("./src/svg.js");
 const PLUGIN_NAME = "remark-mermaid-dataurl";
 
 /**
+ * Converts CLI args to {@link parseMMD} options.
+ *
+ * Required for backwards compatibility.
+ *
+ * @param {{[key: string]: any}} kwargs Args passed to mmdc in format `--key value`
+ * @returns {Promise<import("@mermaid-js/mermaid-cli").ParseMDDOptions>} Options to pass to parseMMD.
+ */
+async function convertMermaidKwargsToParseMMDOpts({
+  theme,
+  width = 800,
+  height = 600,
+  backgroundColor,
+  configFile,
+  cssFile,
+  scale,
+  pdfFit,
+}) {
+  let mermaidConfig = { theme };
+  if (configFile && typeof configFile !== "object") {
+    mermaidConfig = {
+      ...mermaidConfig,
+      ...JSON.parse(await readFile(configFile, { encoding: "utf8" })),
+    };
+  }
+
+  let myCSS;
+  if (cssFile) {
+    myCSS = await readFile(cssFile, { encoding: "utf8" });
+  }
+
+  return {
+    mermaidConfig,
+    backgroundColor,
+    myCSS,
+    pdfFit,
+    viewport: { width, height, deviceScaleFactor: scale },
+  };
+}
+
+/**
  * Calls mmdc (mermaid-cli) with the given keyword args.
  *
  * This function runs `mmdc` in a separate process.
@@ -22,12 +62,6 @@ const PLUGIN_NAME = "remark-mermaid-dataurl";
  * @returns {Promise<string>} Returns the rendered mermaid code as an SVG.
  */
 async function renderMermaidFile(kwargs, input) {
-  let configFile = kwargs.configFile ?? {};
-  if (kwargs.configFile && typeof kwargs.configFile !== "object") {
-    configFile = JSON.parse(
-      await readFile(kwargs.configFile, { encoding: "utf8" })
-    );
-  }
   let puppeteerConfigFile = kwargs.puppeteerConfigFile ?? {};
   if (
     kwargs.puppeteerConfigFile &&
@@ -42,10 +76,12 @@ async function renderMermaidFile(kwargs, input) {
   const { parseMMD } = await import("@mermaid-js/mermaid-cli");
   const browser = await puppeteer.launch(puppeteerConfigFile);
   try {
-    const outputSvg = await parseMMD(browser, input, "svg", {
-      mermaidConfig: configFile,
-      viewport: { width: 800, height: 600 },
-    });
+    const outputSvg = await parseMMD(
+      browser,
+      input,
+      "svg",
+      await convertMermaidKwargsToParseMMDOpts(kwargs)
+    );
     return outputSvg.toString("utf8");
   } finally {
     await browser.close();