+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed some `mermaidCli` options being ignored in v2.0.2
+
 ## [2.0.2] - 2022-08-23
 
 ### Fixed